% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{bonjour Function}
\usage{
bonjour(FirstName = " ")
}
\arguments{
\item{FirstName}{A character string corresponding to
the firstname of the person to wil recieve the hello}
}
\value{
un bonjour qui ne coute rien.
}
\description{
bonjour Function
}
\examples{
bonjour()
}
