% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello Function}
\usage{
hello(FirstName = " ")
}
\arguments{
\item{FirstName}{A character string corresponding to
the firstname of the person to wil recieve the hello}
}
\value{
un bonjour qui ne coute rien.
}
\description{
Hello Function
}
\examples{
hello()

}
