% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importation.r
\name{ImportationCSV}
\alias{ImportationCSV}
\title{Importation of a csv file}
\usage{
ImportationCSV(CSVFilePath)
}
\arguments{
\item{CSVFilePath}{the path of the CSV file.}
}
\value{
DS1 a dataframe with the information contained in the table.
}
\description{
\code{ImportationCSV} returns a dataframe with the information contained in the table.
}
\details{
This function imports a table in a CSV file into a R dataframe.
}
