% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tttestbootstrap.r
\name{ttestbootstrap}
\alias{ttestbootstrap}
\title{ttestbootstrap Function}
\usage{
ttestbootstrap(vGroup1, vGroup2, nbootstrap = 1e+05)
}
\arguments{
\item{vGroup1}{A numerical vector corresponding to the values of the first group}

\item{vGroup2}{A numerical vector corresponding to the values of the second group}

\item{nbootstrap}{The number of bootstrap simulation should be > 100 x (nGroup1 + nGroup1)}
}
\value{
the pvalue calculated by the a simulation of the null hypothesis distribution by bootstrap
}
\description{
ttestbootstrap Function
}
\examples{
ttestbootstrap(vGroup1=rnorm(30,0,1), vGroup2 = rnorm(30,2,1), nbootstrap = 1000)

}
